requirements.txt
streamlit==1.39.0
requests==2.32.3
openai==1.51.0
pypdf==4.3.1